requirements.txt
elasticsearch==9.0.1
faiss==1.7.2
fastapi==0.115.12
huggingface_hub==0.32.3
ing==0.1.0
mteb==1.38.21
numpy==2.2.6
pandas==2.2.3
pydantic==2.11.5
python-dotenv==1.1.0
Requests==2.32.3
sentence_transformers==4.1.0
tiktoken==0.9.0
torch==2.7.0+cu126
tqdm==4.67.1
transformers==4.52.4
uvicorn==0.34.2
